chore: 프로바이더 위치 이동
