+++ src/pages/MyGroup.jsx
@@ -2,44 +2,35 @@ import styled from '@emotion/styled';
 import { COLOR } from '@/styles/color';
 import { imgPuzzle } from '@/assets/images';
 import { Image, Header, Text } from '@/components/base';
-import { getChannelList } from '@/api/channel';
-import { useAsync } from '@/hooks';
 import GroupList from '@/components/domain/GroupList';
-import GroupProvider from '@/context/GroupProvider';
-import TILProvider from '@/context/TILProvider';
 
 function MyGroup() {
-  const initialGroups = useAsync(getChannelList, []);
   return (
-    <GroupProvider initialGroups={initialGroups}>
-      <TILProvider>
-        <StyledPageWrapper>
-          <StyledMyGroup>
-            <StyledHeader>
-              <Header strong size={30}>
-                내 그룹
-              </Header>
-              <button>
-                <i className='fa-solid fa-circle-plus'></i>
-                그룹 만들기
-              </button>
-            </StyledHeader>
-            <StyledDesc>
-              <div>
-                <Text paragraph color={COLOR.DARK} size={2.1} weight={500}>
-                  내가 <Text color={COLOR.TAG_COLOR[0]}>가입한 그룹의 목록</Text>입니다.
-                </Text>
-                <Text paragraph color={COLOR.DARK} size={2.1} weight={500}>
-                  그룹 내의 멤버들과 소통하며 <Text color={COLOR.TAG_COLOR[1]}>오늘의 TIL</Text>을 작성해보세요.
-                </Text>
-              </div>
-              <Image src={imgPuzzle} width={30} />
-            </StyledDesc>
-            <GroupList />
-          </StyledMyGroup>
-        </StyledPageWrapper>
-      </TILProvider>
-    </GroupProvider>
+    <StyledPageWrapper>
+      <StyledMyGroup>
+        <StyledHeader>
+          <Header strong size={30}>
+            내 그룹
+          </Header>
+          <button>
+            <i className='fa-solid fa-circle-plus'></i>
+            그룹 만들기
+          </button>
+        </StyledHeader>
+        <StyledDesc>
+          <div>
+            <Text paragraph color={COLOR.DARK} size={2.1} weight={500}>
+              내가 <Text color={COLOR.TAG_COLOR[0]}>가입한 그룹의 목록</Text>입니다.
+            </Text>
+            <Text paragraph color={COLOR.DARK} size={2.1} weight={500}>
+              그룹 내의 멤버들과 소통하며 <Text color={COLOR.TAG_COLOR[1]}>오늘의 TIL</Text>을 작성해보세요.
+            </Text>
+          </div>
+          <Image src={imgPuzzle} width={30} />
+        </StyledDesc>
+        <GroupList />
+      </StyledMyGroup>
+    </StyledPageWrapper>
   );
 }
